feat: revamp sticker system, command system and add offline support
You can now toggle between checking unread messages as soon as the bot is ready, or not, in the config.json. I have also combined the config.json and lang.json files, so it's easier to edit and manage them. The command system now is very easy to use and does not require messing with internal code.